+++ src/frontend/js/types/api.ts
@@ -1,6 +1,8 @@
 import { Course } from 'types/Course';
 import { FacetedFilterDefinition } from 'types/filters';
-import { Maybe } from 'utils/types';
+import { Maybe, Nullable } from 'utils/types';
+import { User } from 'types/User';
+import { Enrollment } from 'types';
 
 export enum RequestStatus {
   FAILURE = 'failure',
@@ -33,4 +35,32 @@ export interface APICourseSearchResponse {
   };
   meta: APIResponseListMeta;
   objects: Course[];
+}
+
+export interface ApiImplementation {
+  user: {
+    me: () => Promise<Nullable<User>>;
+    login: () => void;
+    register: () => void;
+    logout: () => Promise<void>;
+  };
+  enrollment: {
+    get: (url: string, user: Nullable<User>) => Promise<Nullable<Enrollment>>;
+    isEnrolled: (url: string, user: Nullable<User>) => Promise<boolean>;
+    set: (url: string, user: User) => Promise<boolean>;
+  };
+}
+
+export interface ApiOptions {
+  routes: {
+    [Model: string]: {
+      [Method: string]: string;
+    };
+  };
+}
+
+export enum ApiBackend {
+  BASE = 'base',
+  OPENEDX_DOGWOOD = 'openedx-dogwood',
+  OPENEDX_HAWTHORN = 'openedx-hawthorn',
 }