+++ src/frontend/js/utils/api/authentication.ts
@@ -8,9 +8,10 @@
 import { handle } from 'utils/errors/handle';
 import { AuthenticationBackend } from 'types/commonDataProps';
 import { Nullable } from 'utils/types';
-import { ApiImplementation } from './lms';
+import { ApiImplementation, ApiBackend } from 'types/api';
 import BaseApiInterface from './lms/base';
-import OpenEdxApiInterface from './lms/edx';
+import OpenEdxDogwoodApiInterface from './lms/openedx-dogwood';
+import OpenEdxHawthornApiInterface from './lms/openedx-hawthorn';
 
 const AuthenticationAPIHandler = (): Nullable<ApiImplementation['user']> => {
   const AUTHENTICATION: AuthenticationBackend = (window as any).__richie_frontend_context__?.context
@@ -18,10 +19,12 @@ const AuthenticationAPIHandler = (): Nullable<ApiImplementation['user']> => {
   if (!AUTHENTICATION) return null;
 
   switch (AUTHENTICATION.backend) {
-    case 'richie.apps.courses.lms.base.BaseLMSBackend':
+    case ApiBackend.BASE:
       return BaseApiInterface(AUTHENTICATION).user;
-    case 'richie.apps.courses.lms.edx.TokenEdXLMSBackend':
-      return OpenEdxApiInterface(AUTHENTICATION).user;
+    case ApiBackend.OPENEDX_DOGWOOD:
+      return OpenEdxDogwoodApiInterface(AUTHENTICATION).user;
+    case ApiBackend.OPENEDX_HAWTHORN:
+      return OpenEdxHawthornApiInterface(AUTHENTICATION).user;
     default:
       handle(new Error(`No Authentication Backend found for ${AUTHENTICATION.backend}.`));
       return null;