+++ src/frontend/js/utils/test/factories.ts
@@ -1,5 +1,6 @@
 import { createSpec, derived, faker } from '@helpscout/helix';
 import { CommonDataProps } from 'types/commonDataProps';
+import { ApiBackend } from 'types/api';
 
 const CourseStateFactory = createSpec({
   priority: derived(() => Math.floor(Math.random() * 7)),
@@ -37,12 +38,12 @@ export const ContextFactory = (context: Partial<CommonDataProps['context']> = {}
     csrftoken: faker.random.alphaNumeric(64),
     environment: 'test',
     authentication: {
-      backend: 'richie.apps.courses.lms.base.BaseLMSBackend',
+      backend: ApiBackend.BASE,
       endpoint: 'https://endpoint.test',
     },
     lms_backends: [
       {
-        backend: 'richie.apps.courses.lms.base.BaseLMSBackend',
+        backend: ApiBackend.BASE,
         selector_regexp: '.*',
         course_regexp: '.*',
         endpoint: 'https://endpoint.test',