+++ tests/apps/courses/test_templates_course_detail.py
@@ -470,6 +470,7 @@ class RunsCourseCMSTestCase(CMSTestCase):
         LMS_BACKENDS=[
             {
                 "BACKEND": "richie.apps.courses.lms.edx.TokenEdXLMSBackend",
+                "JS_BACKEND": "openedx-hawthorn",
                 "SELECTOR_REGEX": r".*",
                 "JS_COURSE_REGEX": r"^.*/courses/(?<course_id>.*)/course/?$",
                 "JS_SELECTOR_REGEX": r".*",
@@ -536,6 +537,7 @@ class RunsCourseCMSTestCase(CMSTestCase):
         LMS_BACKENDS=[
             {
                 "BACKEND": "richie.apps.courses.lms.edx.TokenEdXLMSBackend",
+                "JS_BACKEND": "openedx-hawthorn",
                 "SELECTOR_REGEX": r".*",
                 "JS_COURSE_REGEX": r"^.*/courses/(?<course_id>.*)/course/?$",
                 "JS_SELECTOR_REGEX": r".*",