+++ tests/apps/courses/test_templates_course_run_detail.py
@@ -323,6 +323,7 @@ class CourseRunCMSTestCase(CMSTestCase):
         LMS_BACKENDS=[
             {
                 "BACKEND": "richie.apps.courses.lms.edx.TokenEdXLMSBackend",
+                "JS_BACKEND": "openedx-hawthorn",
                 "SELECTOR_REGEX": r".*",
                 "JS_COURSE_REGEX": r"^.*/courses/(?<course_id>.*)/course/?$",
                 "JS_SELECTOR_REGEX": r".*",
@@ -353,6 +354,7 @@ class CourseRunCMSTestCase(CMSTestCase):
         LMS_BACKENDS=[
             {
                 "BACKEND": "richie.apps.courses.lms.edx.TokenEdXLMSBackend",
+                "JS_BACKEND": "openedx-hawthorn",
                 "SELECTOR_REGEX": r".*",
                 "JS_COURSE_REGEX": r"^.*/courses/(?<course_id>.*)/course/?$",
                 "JS_SELECTOR_REGEX": r".*",